i2c: minor cleanup and style changes
