--- drivers/i2c/odroidc4/i2c-client.c
@@ -1,16 +1,0 @@
-#include <microkit.h>
-
-#include "i2c.h"
-#include "printf.h"
-#include "i2c-transport.h"
-
-
-
-void init(void) {
-    sel4cp_dbg_puts("I2C client init\n");
-}
-
-
-void notified(sel4cp_channel c) {
-
-}
