--- drivers/i2c/odroidc4/i2c-sel4cp.c
@@ -1,11 +1,0 @@
-/*
- * Copyright 2023, UNSW
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-// i2c-sel4cp.c
-// Client-server interaction for sDDF i2c driver.
-// Matt Rossouw (matthew.rossouw@unsw.edu.au)
-// 09/2023
-
