+++ drivers/i2c/odroidc4/i2c-transport.c
@@ -22,8 +22,8 @@ uintptr_t ret_free;
 uintptr_t ret_used;
 uintptr_t driver_bufs;
 
-ring_handle_t reqRing;
-ring_handle_t retRing;
+ring_handle_t req_ring;
+ring_handle_t ret_ring;
 
 
 void _putchar(char character) {
@@ -38,8 +38,8 @@ void i2cTransportInit(int buffer_init) {
         microkit_dbg_puts("Not initialising buffers\n");
     }
     // Initialise rings
-    ring_init(&reqRing, (ring_buffer_t *) req_free, (ring_buffer_t *) req_used, buffer_init);
-    ring_init(&retRing, (ring_buffer_t *) ret_free, (ring_buffer_t *) ret_used, buffer_init);
+    ring_init(&req_ring, (ring_buffer_t *) req_free, (ring_buffer_t *) req_used, buffer_init);
+    ring_init(&ret_ring, (ring_buffer_t *) ret_free, (ring_buffer_t *) ret_used, buffer_init);
 
     // If the caller is initialising, also populate the free buffers.
     // Since all buffers are back to back in memory, need to offset each ring's buffers
@@ -48,8 +48,8 @@ void i2cTransportInit(int buffer_init) {
         for (int i = 0; i < I2C_BUF_COUNT; i++) {
             // TODO: check buffer offsetting here. This is definitely too sparse because I haven't updated
             //       it form the 4-buf design
-            enqueue_free(&reqRing, (uintptr_t) driver_bufs + (i * I2C_BUF_SZ), I2C_BUF_SZ);
-            enqueue_free(&retRing, (uintptr_t) driver_bufs + (I2C_BUF_SZ * (i + I2C_BUF_COUNT)), I2C_BUF_SZ);
+            enqueue_free(&req_ring, (uintptr_t) driver_bufs + (i * I2C_BUF_SZ), I2C_BUF_SZ);
+            enqueue_free(&ret_ring, (uintptr_t) driver_bufs + (I2C_BUF_SZ * (i + I2C_BUF_COUNT)), I2C_BUF_SZ);
         }
     }
 
@@ -58,7 +58,7 @@ void i2cTransportInit(int buffer_init) {
 
 req_buf_ptr_t allocReqBuf(size_t size, uint8_t *data, uint8_t client, uint8_t addr) {
     // microkit_dbg_puts("transport: Allocating request buffer\n")
-    if (size > I2C_BUF_SZ - REQ_BUF_DAT_OFFSET*sizeof(i2c_token_t)) {
+    if (size > I2C_BUF_SZ - REQ_BUF_DATA_OFFSET*sizeof(i2c_token_t)) {
         printf("transport: Requested buffer size %zu too large\n", size);
         return 0;
     }
@@ -66,7 +66,7 @@ req_buf_ptr_t allocReqBuf(size_t size, uint8_t *data, uint8_t client, uint8_t ad
     // Allocate a buffer from the appropriate ring
     uintptr_t buf;
     unsigned int sz;
-    int ret = dequeue_free(&reqRing, &buf, &sz);
+    int ret = dequeue_free(&req_ring, &buf, &sz);
     if (ret != 0) {
         return 0;
     }
@@ -74,17 +74,17 @@ req_buf_ptr_t allocReqBuf(size_t size, uint8_t *data, uint8_t client, uint8_t ad
     // Load the client ID and i2c address into first two bytes of buffer
     ((uint8_t *) buf)[REQ_BUF_CLIENT] = client;
     ((uint8_t *) buf)[REQ_BUF_ADDR] = addr;
-    const uint8_t sz_offset = REQ_BUF_DAT_OFFSET*sizeof(uint8_t);
+    const uint8_t sz_offset = REQ_BUF_DATA_OFFSET*sizeof(uint8_t);
 
 
     // Copy the data into the buffer
     memcpy((void *) buf + sz_offset, data, size);
 
     // Enqueue the buffer
-    ret = enqueue_used(&reqRing, buf, size + sz_offset);
+    ret = enqueue_used(&req_ring, buf, size + sz_offset);
     printf("transport: Allocated request buffer %p storing %u bytes\n", buf, size);
     if (ret != 0) {
-        enqueue_free(&reqRing, buf, I2C_BUF_SZ);
+        enqueue_free(&req_ring, buf, I2C_BUF_SZ);
         return 0;
     }
 
@@ -97,7 +97,7 @@ ret_buf_ptr_t getRetBuf() {
     // Allocate a buffer from the appropriate ring
     uintptr_t buf;
     unsigned int sz;
-    int ret = dequeue_free(&retRing, &buf, &sz);
+    int ret = dequeue_free(&ret_ring, &buf, &sz);
     if (ret != 0) {
         microkit_dbg_puts("transport: Failed to get return buffer due to empty free ring!\n");
         return 0;
@@ -114,7 +114,7 @@ int pushRetBuf(ret_buf_ptr_t buf, size_t size) {
     }
 
     // Enqueue the buffer
-    int ret = enqueue_used(&retRing, (uintptr_t)buf, size);
+    int ret = enqueue_used(&ret_ring, (uintptr_t)buf, size);
     if (ret != 0) {
         return 0;
     }
@@ -133,7 +133,7 @@ req_buf_ptr_t popReqBuf(size_t *size) {
     // microkit_dbg_puts("transport: popping request buffer\n");
 
     // Allocate a buffer from the appropriate ring
-    return (req_buf_ptr_t) popBuf(&reqRing, size);
+    return (req_buf_ptr_t) popBuf(&req_ring, size);
 }
 
 
@@ -141,15 +141,15 @@ ret_buf_ptr_t popRetBuf(size_t *size) {
     // microkit_dbg_puts("transport: popping return buffer\n");
 
     // Allocate a buffer from the appropriate ring
-    return (ret_buf_ptr_t) popBuf(&retRing, size);
+    return (ret_buf_ptr_t) popBuf(&ret_ring, size);
 }
 
 int retBufEmpty(void) {
-    return ring_empty(retRing.used_ring);
+    return ring_empty(ret_ring.used_ring);
 }
 
 int reqBufEmpty(void) {
-    return ring_empty(reqRing.used_ring);
+    return ring_empty(req_ring.used_ring);
 }
 
 int releaseReqBuf(req_buf_ptr_t buf) {
@@ -160,7 +160,7 @@ int releaseReqBuf(req_buf_ptr_t buf) {
     }
 
     // Enqueue the buffer
-    int ret = enqueue_free(&reqRing, (uintptr_t)buf, I2C_BUF_SZ);
+    int ret = enqueue_free(&req_ring, (uintptr_t)buf, I2C_BUF_SZ);
     if (ret != 0) {
         return 0;
     }
@@ -175,7 +175,7 @@ int releaseRetBuf(ret_buf_ptr_t buf) {
     }
 
     // Enqueue the buffer
-    int ret = enqueue_free(&retRing, (uintptr_t)buf, I2C_BUF_SZ);
+    int ret = enqueue_free(&ret_ring, (uintptr_t)buf, I2C_BUF_SZ);
     if (ret != 0) {
         return 0;
     }
